fix wii

+++ Game/header/map.hpp
@@ -45,6 +45,6 @@ class MapData{
         //     _ObjectDatas.push_back(*objectData);
         // }
         void SetObjectData(PlaceData *objectData){
-            _TerrainData[objectData->pos.x / MAP_MAGNIFICATION][objectData->pos.y / MAP_MAGNIFICATION][objectData->pos.z / MAP_MAGNIFICATION] = objectData->object;
+            _TerrainData[(int)(objectData->pos.x / MAP_MAGNIFICATION)][(int)(objectData->pos.y / MAP_MAGNIFICATION)][(int)(objectData->pos.z / MAP_MAGNIFICATION)] = objectData->object;
         }
 };
